Hooked in Location Data Model to Dest Loc Picker
Hooked in the Location Data Model to the Destination Location Picker in
the Commute Creator screen. At a minimum, all that needs to be added to
the Commute Creator screen is the alert scheduling logic and then we
would have the minimum number of variables for the Commute Data Model.
I could potentially get all of this created this weekend.