Fixed the scrollview text auto-layout issue
Fixed the issue with the text in the scroll view not being centered.
Was a weird issue where the scrollview doesn’t update its frame size
right away for some reason, but the map does so I just used that as a
basis for the width of the uilabel subviews added into the scrollview.
Whatev, it’s working ha.